redux: better organization
