+++ src/layouts/admin/user.js
@@ -6,7 +6,7 @@ import IconButton from 'material-ui/IconButton'
 import Menu, { MenuItem } from 'material-ui/Menu'
 import AccountCircleIcon from 'material-ui-icons/AccountCircle'
 
-import { logout } from '../../store/actions/auth'
+import { logout } from '../../redux/actions/auth'
 
 class UserAdminLayout extends Component {
   state = {
